requirements.txt
anyio==3.6.2
async-generator==1.10
asyncpg==0.27.0
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.1
decorator==5.1.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
pytest==7.3.0
pytest-asyncio==0.21.0
pytest-tornasync==0.6.0.post2
pytest-trio==0.8.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
requests==2.28.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a33
sqlmodel==0.0.8
starlette==0.26.1
tornado==6.2
trio==0.22.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
